add stylings to all-products
