+++ public/style.css
@@ -72,9 +72,47 @@ form div {
   cursor: pointer;
 }
 
-/* .navbar-links ul li:hover > ul,
-ul li:focus-within > ul, {
-  visibility: visible;
-  opacity: 1;
-  display: block;
-} */
+.products {
+  text-align: center;
+}
+
+.products-search {
+  display: flex;
+  justify-content: center;
+}
+
+.products-search-filter {
+  margin: 10px;
+}
+
+.products-search-select {
+  margin: 10px;
+}
+
+.products-display {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+.product-card {
+  border: 2px solid black;
+  width: 170px;
+  height: 170px;
+  margin: 10px;
+  text-align: center;
+  background-color: antiquewhite;
+  box-shadow: 2px 9px 14px 2px rgba(0,0,0,0.75);
+  border-radius: 8px;
+}
+
+.product-card h4 {
+  color: black;
+  margin: 3px 0;
+}
+
+.product-card-image {
+  height: 100px;
+  width: 140px;
+  object-fit: contain;
+}
